FinancialAgent/requirements.txt
phidata
python-dotenv
yfinance
duckduckgo-search
fastapi
packaging
uvicorn
groq



